Add exhaustive matching to AasxToolkit
This introduces `ExhaustiveMatching.Analyzer` to automatically check
that switches on types and enums are exhaustive.

This is important in dispatching situations where we can easily avoid
future run-time errors caused by the developer forgetting to handle one
or more cases (*e.g.*, when a new class is added to the hierarchy).

The projects `AasxToolkit` and `AasxToolkit.Tests` had to be converted
to SDK-style projects with `dotnet try-convert` in order to make
`ExhaustiveMatching.Analyzer` work. It is not clear why that was
the case.